+++ addons/pvr.mediaportal.tvserver/src/lib/tsreader/MultiFileReader.h
@@ -56,11 +56,11 @@ class MultiFileReader : public FileReader
     virtual long OpenFile();
     virtual long OpenFile(const std::string& fileName);
     virtual long CloseFile();
-    virtual long Read(unsigned char* pbData, unsigned long lDataLength, unsigned long *dwReadBytes);
+    virtual long Read(unsigned char* pbData, size_t lDataLength, ssize_t *dwReadBytes);
 
     virtual bool IsFileInvalid();
 
-    virtual int64_t SetFilePointer(int64_t llDistanceToMove, unsigned long dwMoveMethod);
+    virtual int64_t SetFilePointer(int64_t llDistanceToMove, int dwMoveMethod);
     virtual int64_t GetFilePointer();
     virtual int64_t GetFileSize();
     virtual int64_t OnChannelChange(void);
